--- lib/HomeScreen/home_screen.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomeScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Column();
-  }
-}
