+++ lib/components/choice_clip.dart
