+++ lib/pages/DealsScreen/deals_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DealsScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('This is Deals Screens'),
+    );
+  }
+}
