+++ lib/pages/WatchListScreen/watchlist_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WatchListScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Text('This is WatchList Screens'),
+    );
+  }
+}
